moved configs, add finetune data extraction script, refactor shared methods between run_model and finetune_model
