--- hyak.config
@@ -1,6 +1,0 @@
-PLATFORM="hyak"
-CHECKPOINT="google/pegasus-billsum"
-MODE="predict"
-TESTFILE="preprocess/data/billsum_clean_test_se3_simple_toy.csv"
-CONCAT="post"
-BATCH_SIZE=64
