--- patas.config
@@ -1,6 +1,0 @@
-PLATFORM="patas"
-CHECKPOINT="google/pegasus-billsum"
-MODE="predict"
-TESTFILE="preprocess/data/billsum_clean_test_se3-t5-512-512.csv"
-CONCAT="post"
-BATCH_SIZE=4
